3dtools/requirements.txt
torch>=1.13.1
flask>=2.2.5
#yaml>=0.2.5
pyyaml>=6.0
pillow>=9.4.0
torchvision>=0.14.1
numpy>=1.22.2
opencv-python>=4.7.0.72
matplotlib>=3.5.3
pandas>=1.3.5
seaborn>=0.12.2
scipy>=1.7.3
ultralytics>=8.0.117
requests>=2.31.0
tqdm>=4.65.0
psutil>=5.9.5
thop>=0.1.1-2209072238
scikit-image>=0.19.3
ultralytics>=8.1.0